requirements.txt
appdirs==1.4.4
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Fabric3==1.14.post1
filelock==3.12.0
flasgger==0.9.5
Flask==2.1.0
Flask-Cors==3.0.10
greenlet==2.0.2
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
install==1.3.5
invoke==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.5.0
MarkupSafe==2.1.2
mistune==2.0.5
more-itertools==4.2.0
mysqlclient==2.1.1
netifaces==0.10.4
oauthlib==3.1.0
paramiko==2.12.0
pbr==5.11.1
pep8==1.7.1
pexpect==4.6.0
platformdirs==3.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.10.0
pycparser==2.21
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
python-http-client==3.3.7
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sendgrid==6.10.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
soupsieve==1.9.5
SQLAlchemy==2.0.7
ssh-import-id==5.10
starkbank-ecdsa==2.2.0
systemd-python==234
testresources==2.0.1
tomli==2.0.1
Twisted==18.9.0
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.22.0
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.2.3
zipp==1.0.0
zope.interface==4.7.1
